refactor(analysis): move source complexity tests to owner module
